shope/requirements.txt
Django==4.2
Pillow==9.3.0
flake8==6.0.0
pre-commit==3.2.2
python-dotenv==1.0.0
django-phonenumber-field==7.0.2
django-taggit==3.1.0
pytz==2023.3
requests==2.29.0
pymemcache==4.0.0
celery==5.2.7
redis==4.5.5
django-silk==5.0.3
gprof2dot==2022.7.29
pandas==2.0.2
openpyxl==3.1.2
XlsxWriter==3.1.2
xlrd==2.0.1
psycopg2-binary==2.9.5
phonenumbers==8.13.11
inject==5.0.0
yookassa==2.4.0
beartype==0.15.0
django_extensions==3.2.3
djangorestframework==3.14.0
django-debug-toolbar==4.2.0